// src/gen-ts/src/models/ExpenseMoney.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { CrupdateJob } from './CrupdateJob'
import {
  CrupdateJobFromJSON,
  CrupdateJobFromJSONTyped,
  CrupdateJobToJSON,
  CrupdateJobToJSONTyped,
} from './CrupdateJob'
import type { AuditUser } from './AuditUser'
import {
  AuditUserFromJSON,
  AuditUserFromJSONTyped,
  AuditUserToJSON,
  AuditUserToJSONTyped,
} from './AuditUser'

/**
 *
 * @export
 * @interface ExpenseMoney
 */
export interface ExpenseMoney {
  /**
   *
   * @type {string}
   * @memberof ExpenseMoney
   */
  comment?: string
  /**
   *
   * @type {Date}
   * @memberof ExpenseMoney
   */
  created_at?: Date
  /**
   *
   * @type {Date}
   * @memberof ExpenseMoney
   */
  updated_at?: Date
  /**
   *
   * @type {AuditUser}
   * @memberof ExpenseMoney
   */
  created_by?: AuditUser
  /**
   *
   * @type {AuditUser}
   * @memberof ExpenseMoney
   */
  updated_by?: AuditUser
  /**
   *
   * @type {number}
   * @memberof ExpenseMoney
   */
  amount?: number
  /**
   *
   * @type {string}
   * @memberof ExpenseMoney
   */
  description?: string
  /**
   *
   * @type {string}
   * @memberof ExpenseMoney
   */
  id?: string
  /**
   * Job this expense is related to
   * @type {CrupdateJob}
   * @memberof ExpenseMoney
   */
  job?: CrupdateJob
}

/**
 * Check if a given object implements the ExpenseMoney interface.
 */
export function instanceOfExpenseMoney(value: object): value is ExpenseMoney {
  return true
}

export function ExpenseMoneyFromJSON(json: any): ExpenseMoney {
  return ExpenseMoneyFromJSONTyped(json, false)
}

export function ExpenseMoneyFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExpenseMoney {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    created_at: json['created_at'] == null ? undefined : new Date(json['created_at']),
    updated_at: json['updated_at'] == null ? undefined : new Date(json['updated_at']),
    created_by: json['created_by'] == null ? undefined : AuditUserFromJSON(json['created_by']),
    updated_by: json['updated_by'] == null ? undefined : AuditUserFromJSON(json['updated_by']),
    amount: json['amount'] == null ? undefined : json['amount'],
    description: json['description'] == null ? undefined : json['description'],
    id: json['id'] == null ? undefined : json['id'],
    job: json['job'] == null ? undefined : CrupdateJobFromJSON(json['job']),
  }
}

export function ExpenseMoneyToJSON(json: any): ExpenseMoney {
  return ExpenseMoneyToJSONTyped(json, false)
}

export function ExpenseMoneyToJSONTyped(
  value?: ExpenseMoney | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    created_at:
      value['created_at'] == null ? value['created_at'] : value['created_at'].toISOString(),
    updated_at:
      value['updated_at'] == null ? value['updated_at'] : value['updated_at'].toISOString(),
    created_by: AuditUserToJSON(value['created_by']),
    updated_by: AuditUserToJSON(value['updated_by']),
    amount: value['amount'],
    description: value['description'],
    id: value['id'],
    job: CrupdateJobToJSON(value['job']),
  }
}